requirements.txt
streamlit
chromadb
langchain
langchain-core
langchain-community
langchain-google-genai
sentence-transformers
unstructured[local-inference]
pymupdf
python-pptx
google-generativeai
pillow